+++ src/tests/utils.ts
@@ -459,6 +459,7 @@ export const externalAddressApiMock = {
   listExternalAddressBalances: jest.fn(),
   getExternalAddressBalance: jest.fn(),
   requestExternalFaucetFunds: jest.fn(),
+  listAddressHistoricalBalance: jest.fn(),
 };
 
 export const serverSignersApiMock = {